♻️ Extract the common logic into a single base module.
