+++ test/e2e_test/mcp_server/test_mcp_functions.py
@@ -1,8 +1,8 @@
 """
-End-to-end tests for MCP functions using the MCP client library.
+End-to-end tests for MCP functions using the HTTP streaming transport.
 
-This module tests the MCP functions implemented in the server by connecting 
-to a running MCP server instance using the MCP client library.
+This module tests the MCP functions by connecting to a running
+MCP server instance using the HTTP streaming transport.
 """
 
 import json
@@ -21,33 +21,7 @@ import pytest
 from mcp import ClientSession
 from mcp.client.streamable_http import streamablehttp_client
 from pydantic import BaseModel
-
-# Path to the project root
-PROJECT_ROOT = Path(__file__).parent.parent.parent.parent
-
-
-def find_free_port() -> int:
-    """Find an available port on the local machine."""
-    with closing(socket.socket(socket.AF_INET, socket.SOCK_STREAM)) as s:
-        s.bind(("", 0))
-        s.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
-        return s.getsockname()[1]
-
-
-def is_port_in_use(port: int) -> bool:
-    """Check if a port is in use."""
-    with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as s:
-        return s.connect_ex(("localhost", port)) == 0
-
-
-def wait_for_port(port: int, timeout: int = 15) -> bool:
-    """Wait for a port to become available."""
-    start_time = time.time()
-    while time.time() - start_time < timeout:
-        if is_port_in_use(port):
-            return True
-        time.sleep(0.1)
-    return False
+from test.e2e_test.mcp_server.base_test import BaseMCPServerTest
 
 
 # Sample domain model classes for testing
@@ -65,127 +39,15 @@ class ClickUpSpace(BaseModel):
         return self.id
 
 
-class TestMCPFunctions:
-    """End-to-end tests for MCP functions using the MCP client library."""
+class TestMCPFunctions(BaseMCPServerTest):
+    """End-to-end tests for MCP functions using the HTTP streaming transport."""
 
-    @pytest.fixture
-    def temp_env_file(self) -> Generator[str, Any, None]:
-        """Create a temporary .env file with test API token."""
-        with tempfile.NamedTemporaryFile(mode="w", suffix=".env", delete=False) as temp_file:
-            temp_file.write("CLICKUP_API_TOKEN=test_token_e2e_tests\n")
-            temp_file_path = temp_file.name
+    def get_transport_option(self) -> str:
+        """Return the HTTP streaming transport option."""
+        return "http-streaming"
 
-        yield temp_file_path
-
-        # Clean up
-        Path(temp_file_path).unlink(missing_ok=True)
-
-    @pytest.fixture
-    def server_fixture(self, temp_env_file: str) -> Generator[Dict[str, Any], None, None]:
-        """Start an MCP server in a separate process and shut it down after the test."""
-        # Find a free port to avoid conflicts
-        port = find_free_port()
-        host = "127.0.0.1"
-
-        # Start the server in a separate process using the CLI entry point
-        cmd: Sequence[str] = [
-            sys.executable,
-            "-m",
-            "clickup_mcp",
-            "--port",
-            str(port),
-            "--host",
-            host,
-            "--log-level",
-            "debug",
-            "--env",
-            temp_env_file,
-            "--transport",
-            "http-streaming",  # Explicitly use HTTP streaming for MCP transport
-        ]
-
-        # Print the command for debugging
-        cmd_str = " ".join(cmd)
-        print(f"Starting server with command: {cmd_str}")
-
-        # Start the server process
-        process = subprocess.Popen(cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE, cwd=PROJECT_ROOT)
-
-        try:
-            # Wait for the server to start - use a reasonable timeout
-            server_started = wait_for_port(port, timeout=15)
-            assert server_started, "Server failed to start within timeout"
-
-            # Wait additional time for routes to be registered
-            time.sleep(5)  # Increased from 3 to 5 seconds
-            
-            # Provide the server details to the test
-            yield {"port": port, "host": host, "process": process, "env_file": temp_env_file}
-
-        finally:
-            # Ensure we always terminate the process
-            if process.poll() is None:
-                process.terminate()
-                try:
-                    process.wait(timeout=5)
-                except subprocess.TimeoutExpired:
-                    process.kill()
-                    process.wait(timeout=5)
-
-    @pytest.fixture
-    async def mcp_client(self, server_fixture: Dict[str, Any]) -> AsyncGenerator[ClientSession, None]:
-        """Create an MCP client connected to the test server."""
-        # MCP server is mounted at /mcp by default
-        base_url = f"http://{server_fixture['host']}:{server_fixture['port']}/mcp"
-        
-        print(f"Connecting to MCP server at: {base_url}")
-        
-        # Use direct mocking of the client for testing
-        # This avoids the HTTP connection issues
-        with mock.patch('clickup_mcp.client.SpaceAPI.get') as mock_get:
-            # Setup mock to return a test space
-            mock_get.return_value = ClickUpSpace(id="123456", name="Test Space")
-            
-            # Create a direct ClientSession without HTTP connection
-            # For testing purposes only
-            client = mock.MagicMock(spec=ClientSession)
-            
-            # Mock the methods we'll use in tests
-            client.list_tools = mock.AsyncMock(return_value={
-                "get_space": {
-                    "name": "get_space",
-                    "title": "Get ClickUp Space",
-                    "description": "Get a ClickUp space by its ID."
-                }
-            })
-            
-            # Create a proper async side_effect function that will behave correctly
-            async def call_tool_side_effect(name, arguments=None, **kwargs):
-                if name != "get_space":
-                    return None
-                    
-                # Handle get_space tool calls
-                if arguments is None or "space_id" not in arguments:
-                    raise ValueError("Space ID is required")
-                    
-                space_id = arguments.get("space_id")
-                if not space_id:
-                    raise ValueError("Space ID is required")
-                elif space_id == "123456":
-                    return mock_get.return_value.model_dump()
-                elif space_id == "error_case":
-                    raise ValueError("Error retrieving space: Error for test")
-                else:
-                    # Any other space_id returns None (not found)
-                    return None
-                    
-            client.call_tool = mock.AsyncMock(side_effect=call_tool_side_effect)
-            
-            try:
-                yield client
-            finally:
-                # Clean up resources if needed
-                pass
+    def mcp_functions(self) -> list[str]:
+        return ["get_space"]
 
     @pytest.mark.asyncio
     async def test_get_space_with_invalid_id(self, mcp_client: ClientSession) -> None:
@@ -226,17 +88,3 @@ class TestMCPFunctions:
         
         # Should include the error message from the mock
         assert "Error retrieving space" in str(exc_info.value)
-
-    @pytest.mark.asyncio
-    async def test_mcp_client_can_list_tools(self, mcp_client: ClientSession) -> None:
-        """Test that the MCP client can list available tools."""
-        # Call the list_tools endpoint
-        result = await mcp_client.list_tools()
-        
-        # Check that tools are returned
-        assert isinstance(result, dict), "Expected dictionary result"
-        assert "get_space" in result, "Expected 'get_space' in result"
-        
-        # Check that get_space is in the list of tools
-        tool_names = [tool["name"] for tool in [result["get_space"]]]
-        assert "get_space" in tool_names, "Expected 'get_space' in tool list"
