+++ test/e2e_test/mcp_server/test_stdio_client.py
@@ -20,32 +20,7 @@ import pytest
 from mcp import ClientSession
 from pydantic import BaseModel
 
-# Path to the project root
-PROJECT_ROOT = Path(__file__).parent.parent.parent.parent
-
-
-def find_free_port() -> int:
-    """Find an available port on the local machine."""
-    with closing(socket.socket(socket.AF_INET, socket.SOCK_STREAM)) as s:
-        s.bind(("", 0))
-        s.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
-        return s.getsockname()[1]
-
-
-def is_port_in_use(port: int) -> bool:
-    """Check if a port is in use."""
-    with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as s:
-        return s.connect_ex(("localhost", port)) == 0
-
-
-def wait_for_port(port: int, timeout: int = 15) -> bool:
-    """Wait for a port to become available."""
-    start_time = time.time()
-    while time.time() - start_time < timeout:
-        if is_port_in_use(port):
-            return True
-        time.sleep(0.1)
-    return False
+from test.e2e_test.mcp_server.base_test import BaseMCPServerTest
 
 
 # Sample domain model classes for testing
@@ -62,132 +37,22 @@ class ClickUpSpace(BaseModel):
         return self.id
 
 
-class TestSSETransport:
+class TestSSETransport(BaseMCPServerTest):
     """End-to-end tests for MCP functions using the SSE transport."""
 
-    @pytest.fixture
-    def temp_env_file(self) -> Generator[str, Any, None]:
-        """Create a temporary .env file with test API token."""
-        with tempfile.NamedTemporaryFile(mode="w", suffix=".env", delete=False) as temp_file:
-            temp_file.write("CLICKUP_API_TOKEN=test_token_e2e_tests\n")
-            temp_file_path = temp_file.name
+    def get_transport_option(self) -> str:
+        """Return the SSE transport option."""
+        return "sse"
 
-        yield temp_file_path
-
-        # Clean up
-        Path(temp_file_path).unlink(missing_ok=True)
-
-    @pytest.fixture
-    def server_fixture(self, temp_env_file: str) -> Generator[Dict[str, Any], None, None]:
-        """Start an MCP server in a separate process and shut it down after the test."""
-        # Find a free port to avoid conflicts
-        port = find_free_port()
-        host = "127.0.0.1"
-
-        # Start the server in a separate process using the CLI entry point
-        cmd: Sequence[str] = [
-            sys.executable,
-            "-m",
-            "clickup_mcp",
-            "--port",
-            str(port),
-            "--host",
-            host,
-            "--log-level",
-            "debug",
-            "--env",
-            temp_env_file,
-            "--transport",
-            "sse",  # Use SSE transport instead of http-streaming
-        ]
-
-        # Print the command for debugging
-        cmd_str = " ".join(cmd)
-        print(f"Starting server with command: {cmd_str}")
-
-        # Start the server process
-        process = subprocess.Popen(cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE, cwd=PROJECT_ROOT)
-
-        try:
-            # Wait for the server to start - use a reasonable timeout
-            server_started = wait_for_port(port, timeout=15)
-            assert server_started, "Server failed to start within timeout"
-
-            # Wait additional time for routes to be registered
-            time.sleep(5)  # Increased from 3 to 5 seconds
-            
-            # Provide the server details to the test
-            yield {"port": port, "host": host, "process": process, "env_file": temp_env_file}
-
-        finally:
-            # Ensure we always terminate the process
-            if process.poll() is None:
-                process.terminate()
-                try:
-                    process.wait(timeout=5)
-                except subprocess.TimeoutExpired:
-                    process.kill()
-                    process.wait(timeout=5)
-
-    @pytest.fixture
-    async def mcp_client(self, server_fixture: Dict[str, Any]) -> AsyncGenerator[ClientSession, None]:
-        """Create an MCP client connected to the test server."""
-        # MCP server is mounted at /mcp by default
-        base_url = f"http://{server_fixture['host']}:{server_fixture['port']}/mcp"
-        
-        print(f"Connecting to MCP server at: {base_url}")
-        
-        # Use direct mocking of the client for testing
-        with mock.patch('clickup_mcp.client.SpaceAPI.get') as mock_get:
-            # Setup mock to return a test space
-            mock_get.return_value = ClickUpSpace(id="123456", name="Test Space")
-            
-            # Create a direct ClientSession without HTTP connection
-            client = mock.MagicMock(spec=ClientSession)
-            
-            # Mock the methods we'll use in tests
-            client.list_tools = mock.AsyncMock(return_value={
-                "get_space": {
-                    "name": "get_space",
-                    "title": "Get ClickUp Space",
-                    "description": "Get a ClickUp space by its ID."
-                }
-            })
-            
-            # Create a proper async side_effect function that will behave correctly
-            async def call_tool_side_effect(name, arguments=None, **kwargs):
-                if name != "get_space":
-                    return None
-                    
-                # Handle get_space tool calls
-                if arguments is None or "space_id" not in arguments:
-                    raise ValueError("Space ID is required")
-                    
-                space_id = arguments.get("space_id")
-                if not space_id:
-                    raise ValueError("Space ID is required")
-                elif space_id == "123456":
-                    return mock_get.return_value.model_dump()
-                elif space_id == "error_case":
-                    raise ValueError("Error retrieving space: Error for test")
-                else:
-                    # Any other space_id returns None (not found)
-                    return None
-                    
-            client.call_tool = mock.AsyncMock(side_effect=call_tool_side_effect)
-            
-            try:
-                yield client
-            finally:
-                # Clean up resources if needed
-                pass
+    def mcp_functions(self) -> list[str]:
+        return ["get_space"]
 
     @pytest.mark.asyncio
     async def test_get_space_with_invalid_id(self, mcp_client: ClientSession) -> None:
         """Test that get_space returns None for an invalid space ID."""
         # Call the MCP function with an invalid space ID
         result = await mcp_client.call_tool("get_space", {"space_id": "invalid_id"})
-        
+
         # Should return None for invalid space ID
         assert result is None, "Expected None for invalid space ID"
 
@@ -197,7 +62,7 @@ class TestSSETransport:
         # Call the MCP function with an empty space ID
         with pytest.raises(ValueError) as exc_info:
             await mcp_client.call_tool("get_space", {"space_id": ""})
-        
+
         # Check the error message
         assert "Space ID is required" in str(exc_info.value)
 
@@ -206,7 +71,7 @@ class TestSSETransport:
         """Test get_space with a mocked response."""
         # Call the MCP function with a valid space ID that our mock will handle
         result = await mcp_client.call_tool("get_space", {"space_id": "123456"})
-        
+
         # Check the response matches our mock
         assert result is not None
         assert result["id"] == "123456"
@@ -218,20 +83,6 @@ class TestSSETransport:
         # This will hit our error case in the mock's side_effect
         with pytest.raises(ValueError) as exc_info:
             await mcp_client.call_tool("get_space", {"space_id": "error_case"})
-        
+
         # Should include the error message from the mock
         assert "Error retrieving space" in str(exc_info.value)
-
-    @pytest.mark.asyncio
-    async def test_mcp_client_can_list_tools(self, mcp_client: ClientSession) -> None:
-        """Test that the MCP client can list available tools."""
-        # Call the list_tools endpoint
-        result = await mcp_client.list_tools()
-        
-        # Check that tools are returned
-        assert isinstance(result, dict), "Expected dictionary result"
-        assert "get_space" in result, "Expected 'get_space' in result"
-        
-        # Check that get_space is in the list of tools
-        tool_names = [tool["name"] for tool in [result["get_space"]]]
-        assert "get_space" in tool_names, "Expected 'get_space' in tool list"
